Primer avance de catalogo
Se completó la funcionalidad general del catalogo, falta hacer una mejoras y seccionarlo de mejor manera
