+++ mercado_local_proy/mercado_local_app/templates/inicio.html
@@ -9,4 +9,4 @@
 
     <ul>
 </nav>
-<p>hola</p>
+
