+++ mercado_local_proy/mercado_local_app/templates/layout.html
@@ -41,16 +41,20 @@
         <span class = "relleno-barra"></span>
     </div>
 
+    
+
    
 
     <div id="content">
         <div class="box">
 
         {% block content %}
+        <p>Hola</p>
         {% endblock %}
         </div>
         
     </div>
+
     <footer>
         MercadoLocal &copy; - {% now "Y" %}
     </footer>
