+++ mercado_local_proy/mercado_local_proy/settings.py
@@ -82,7 +82,7 @@ WSGI_APPLICATION = 'mercado_local_proy.wsgi.application'
 DATABASES = {
    'default': {
         'ENGINE': 'django.db.backends.postgresql',
-        "NAME": "mercadoLocal",
+        "NAME": "mercadoL",
         "USER": "postgres",
         "PASSWORD": "root",
         "HOST": "127.0.0.1",
