+++ mercado_local_proy/productos_app/models.py
@@ -1,5 +1,5 @@
 from django.db import models
-from mercado_local_app.models  import Negocio, Cliente
+from mercado_local_app.models  import Negocio
 
 # Create your models here.
 
@@ -12,21 +12,5 @@ class Producto(models.Model):
     categoria = models.CharField(max_length=30)  
     precio = models.FloatField()
 
-class Venta(models.Model):
-    costoTotal = models.IntegerField(null=False, blank=False)
-    cliente = models.ForeignKey(Cliente,on_delete=models.CASCADE )
 
-class Orden(models.Model):
-    producto = models.ForeignKey(Producto, on_delete=models.CASCADE)
-    venta =  models.ForeignKey(Venta, null=True, on_delete=models.CASCADE)
-    cantidadProducto = models.IntegerField(null=False, blank=False)
-    costoOrden = models.IntegerField(null=False, blank=False)
-
-
-
-class HistorialCompra(models.Model):
-    producto = models.ForeignKey(Producto, null=False, on_delete=models.PROTECT)
-    costoCompra = models.IntegerField(null=False, blank=False)
-    venta = models.ForeignKey(Venta, null=True, on_delete=models.SET_NULL)
-    orden = models.ForeignKey(Orden, null=True, on_delete=models.SET_NULL)
