+++ mercado_local_proy/productos_app/templates/productos_app/catalogo_producto.html
@@ -1,0 +1,24 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Document</title>
+</head>
+<body>
+
+    <h1>MERCADO LOCAL</h1>
+
+    {% if catalogo_principal %}
+        <ul>
+            {% for producto in catalogo_principal %}
+            <li><a href="{% url 'detalle_producto' producto.id %}">
+                {{ producto.nombreProducto}}</a></li>
+            {% endfor %}
+        </ul>
+    {% else  %}
+        <p>No hay productos</p>
+    {% endif %}
+ 
+</body>
+</html>
