+++ mercado_local_proy/productos_app/templates/productos_app/detalle_producto.html
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Document</title>
+</head>
+<body>
+    
+    <h1>{{ producto.nombreProducto }}</h1>
+    <p>Contenido: {{ producto.descripcionProducto}}</p>
+
+    <a href="{% url 'inicio' %}">Volver al inicio</a>
+    
+</body>
+</html>
